pyproject.toml
[tool.poetry]
name = "nautobot-chatops"
version = "1.0.0"
description = "A plugin providing chatbot capabilities for Nautobot"
authors = ["Network to Code, LLC <opensource@networktocode.com>"]
readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-nautobot"
repository = "https://github.com/nautobot/nautobot-plugin-nautobot"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "LICENSE",
    "README.md",
]

[tool.poetry.plugins."nautobot.workers"]
"clear" = "nautobot_chatops.workers.clear:clear"
"nautobot" = "nautobot_chatops.workers.nautobot:nautobot"

[tool.poetry.dependencies]
python = "^3.6"
nautobot = "*"
nautobot-capacity-metrics = "*"
slackclient = "^2.7.1"
texttable = "^1.6.2"
pyjwt = {extras = ["crypto"], version = "^1.7.1"}
webexteamssdk = "^1.3"

[tool.poetry.dev-dependencies]
black = "*"
yamllint = "*"
bandit = "*"
pylint = "*"
pylint-django = "*"
pydocstyle = "*"
prybar = "*"
invoke = "*"
flake8 = "*"

[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.masonry.api"

